java練習問題.修正

+++ HowOldAreYou.java
@@ -19,8 +19,9 @@ public class HowOldAreYou {
 					System.out.println(line2+"さんの年齢を入力してください。");
 					line2 = reader.readLine();
 
-					int age = Integer.parseInt(line);
-					int age2 = Integer.parseInt(line2);
+                    //平均年齢なのでintではなくdouble
+					double age = Integer.parseInt(line);
+					double age2 = Integer.parseInt(line2);
 
 					//2人の平均年齢を出す
 					System.out.println("その2人の年齢の平均は"+(age+age2)/2+"です。");
